OpenLog Analog - OpenLog that reads analog ports
This is a knock-off of the SparkFun OpenLog that brings the AVR's analog
ports out to a header and modifies the software to record them at a rate
of 100 samples per second (each port, recorded in a batch).

This first version does polling of a timer to determine the interval between
samples.  This could result in irregular sample intervals due to SD card
writes.